// annix-backend/src/hdpe/dto/pipe-cost-response.dto.ts
export class PipeCostResponseDto {
  nominalBore: number;
  sdr: number;
  length: number;
  outerDiameter: number;
  wallThickness: number;
  innerDiameter: number;
  weightKgPerM: number;
  totalWeight: number;
  numButtwelds: number;
  materialCost: number;
  buttweldCost: number;
  totalCost: number;
  pricePerKg: number;
  buttweldPrice: number;
}